+++ xarray/backends/pydap_.py
@@ -28,12 +28,6 @@ class PydapArrayWrapper(BackendArray):
         # downloading coordinate data twice
         array = getattr(self.array, 'array', self.array)
         result = robust_getitem(array, key, catch=ValueError)
-        # pydap doesn't squeeze axes automatically like numpy
-        axis = tuple(n for n, k in enumerate(key)
-                     if isinstance(k, integer_types))
-        if len(axis) > 0:
-            result = np.squeeze(result, axis)
-
         return result